Getting more precision for ms time

+++ src/Zoxive.HttpLoadTesting.Client/Domain/Database/DbInitializer.cs
@@ -44,7 +44,7 @@ CREATE TABLE HttpStatusResult (
     Id                  INTEGER PRIMARY KEY,
     IterationId         INTEGER REFERENCES Iteration (Id) ON DELETE CASCADE,
     Method              VARCHAR,
-    ElapsedMilliseconds BIGINT,
+    ElapsedMilliseconds REAL, 
     RequestUrl          VARCHAR,
     StatusCode          INTEGER
 );
