+++ src/Zoxive.HttpLoadTesting.Client/Domain/HttpStatusResult/Dtos/HttpStatusResultDto.cs
@@ -12,7 +12,7 @@ namespace Zoxive.HttpLoadTesting.Client.Domain.HttpStatusResult.Dtos
 
         public string Method { get; set; }
 
-        public long ElapsedMilliseconds { get; set; }
+        public double ElapsedMilliseconds { get; set; }
 
         public string RequestUrl { get; set; }
